rearchitect and remove weather state
