+++ lib/pages/main_page.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_training/views/weather_view.dart';
+import 'package:flutter_training/components/weather_view.dart';
 
 class MainPage extends StatelessWidget {
   const MainPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
+    return Scaffold(
       body: WeatherView(),
     );
   }
