+++ lib/providers/fetch_weather_provider.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:flutter_training/models/weather.dart';
 import 'package:flutter_training/models/weather_request.dart';
 import 'package:flutter_training/providers/yumemi_weather_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -7,11 +8,13 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'fetch_weather_provider.g.dart';
 
 @riverpod
-Future<String> fetchWeather(
+Future<Weather> fetchWeather(
   FetchWeatherRef ref, {
   required WeatherRequest weatherRequest,
 }) async {
   final yumemiWeatherClient = ref.watch(yumemiWeatherProvider);
   final jsonText = jsonEncode(weatherRequest.toJson());
-  return yumemiWeatherClient.fetchWeather(jsonText);
+  final json = yumemiWeatherClient.fetchWeather(jsonText);
+  final map = jsonDecode(json) as Map<String, dynamic>;
+  return Weather.fromJson(map);
 }
